separate crd command group
